Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/track/track-element/track-cue-negative-duration-expected.txt



FAIL Enter, Exit events for a cue with negative duration assert_equals: expected 1 but got 0
